+++ src/onegov.activity/onegov/activity/models/invoice_item.py
@@ -2,6 +2,7 @@ from onegov.activity import utils
 from onegov.core.orm import Base
 from onegov.core.orm.mixins import TimestampMixin
 from onegov.core.orm.types import UUID
+from onegov.pay import Payable
 from onegov.user import User
 from sqlalchemy import Boolean
 from sqlalchemy import Column
@@ -14,7 +15,7 @@ from sqlalchemy_utils import observes
 from uuid import uuid4
 
 
-class InvoiceItem(Base, TimestampMixin):
+class InvoiceItem(Base, TimestampMixin, Payable):
     """ An item in an invoice. """
 
     __tablename__ = 'invoice_items'
@@ -64,7 +65,7 @@ class InvoiceItem(Base, TimestampMixin):
 
     @validates('source')
     def validate_source(self, key, value):
-        assert value in (None, 'xml', 'stripe')
+        assert value in (None, 'xml', 'stripe_connect')
         return value
 
     @property